pyproject.toml
[project]
name = "foreign-domains"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiodns>=3.2.0",
    "asyncio>=3.4.3",
    "ipaddress>=1.0.23",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "asyncpg>=0.30.0",
]
